tsubasa: fstab: Allow recovery to use the misc partition
Android Nougat changed how it communicates with the recovery
partition. Until now the recovery commands had been written
to the /cache partition, but this has changed. Now recovery
writes to the /misc partition.

Fortunately for blue devices we actually have
an empty and unused /apps_log partition we can use.

Change-Id: Ie200d65b3e8b532cb9537d302f13ebb8f7c17af4
Signed-off-by: Adam Farden <adam@farden.cz>